renomme le nom du package de motifs à pattern
