--- src/fr/jgetmove/jgetmove/motifs/Pattern.java
@@ -1,5 +1,0 @@
-package fr.jgetmove.jgetmove.motifs;
-
-public interface Pattern {
-
-}
