+++ src/fr/jgetmove/jgetmove/pattern/Pattern.java
@@ -1,0 +1,5 @@
+package fr.jgetmove.jgetmove.pattern;
+
+public interface Pattern {
+
+}
