+++ src/fr/jgetmove/jgetmove/pattern/Swarm.java
@@ -1,4 +1,4 @@
-package fr.jgetmove.jgetmove.motifs;
+package fr.jgetmove.jgetmove.pattern;
 
 public abstract class Swarm implements Pattern {
