+++ src/fr/jgetmove/jgetmove/solver/Generator.java
@@ -6,5 +6,5 @@ import java.util.ArrayList;
 
 public interface Generator {
 
-    public ArrayList<ArrayList<Integer>> generateClusters(Database database);
+    public ArrayList<ArrayList<Integer>> generate(Database database);
 }
